% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA.DBDS.R
\name{performKNN_DensityBasedDownsampling}
\alias{performKNN_DensityBasedDownsampling}
\title{performKNN_DensityBasedDownsampling}
\usage{
performKNN_DensityBasedDownsampling(
  input.data,
  downsample_n = 500,
  features,
  k = 25,
  radius = 0,
  min_value = 1,
  max_value = 500
)
}
\arguments{
\item{input.data}{your dataset}

\item{downsample_n}{# of observations to downsample to}

\item{features}{vector of feautures to use for KNN search. Features must be in the input.data}

\item{k}{# of neighbors used in NN search}

\item{radius}{Radius of search for searchtype='radius'}

\item{min_value}{minimum value used for normalization range}

\item{max_value}{minimum value used for normalization range}
}
\description{
This function performs KNN DBDS
}
\keyword{internal}
